profile

+++ profile/delete.php
@@ -1,9 +1,12 @@
 <?php
   include('../koneksi.php');
 
-  $id = $_SESSION['id'];
-  $sql="DELETE FROM register WHERE id_user='$id'";
-  if (mysqli_query($con,$sql)) {
-    echo "Data Deleted";
+  if($isset($_POST['delete'])){
+    $_SESSION["id"] = $row['id_user'];
+    $sql="DELETE FROM register WHERE id_user='$id'";
+    if (mysqli_query($con,$sql)) {
+      echo "Data Deleted";
+    }
   }
+  
  ?>
